+++ src/components/Popups/Alerts/components/AlertFooter.tsx
@@ -1,11 +1,13 @@
 import React, { useEffect } from "react";
 
+import { Icon } from "../../../Miscellaneous/Icon";
 import { AlertFooterProps } from "../types";
 
 export function AlertFooter<TData = undefined>({
   data,
   confirmBtnText = "Confirm",
   confirmBtnDisabled = false,
+  confirmBtnBusy = false,
   disableConfirmOnEnter = false,
   onConfirm,
   declineBtnText = "Close",
@@ -42,10 +44,13 @@ export function AlertFooter<TData = undefined>({
     <div className="m-alert-footer">
       {onConfirm && (
         <button
-          disabled={confirmBtnDisabled}
+          disabled={confirmBtnDisabled || confirmBtnBusy}
           onClick={() => onConfirm(data as unknown as TData)}
           className="m-alert-confirm-button"
         >
+          {confirmBtnBusy && (
+            <Icon icon={["fas", "circle-notch"]} className="fa-spin m-alert-confirm-button-busy-icon" />
+          )}
           {confirmBtnText}
         </button>
       )}